+++ tests/test_service.py
@@ -16,6 +16,6 @@ class TestService:
         user = api_client.user.get_me()
         project = api_client.project.list(user_id=user.id)[0]
         remote = api_client.remote.list(project.id)[0]
-        service = api_client.service.list(remote_id=remote.id, service_type='ssh')
+        service = api_client.service.list(remote_id=remote.id, service_type='deploy')
 
         assert isinstance(service, list) and hasattr(service[0], 'id') and uuid.UUID(service[0].id)